feat: add Ink The Future section
