+++ src/app/[locale]/about/page.tsx
@@ -4,14 +4,18 @@ import { newLayoutContainerClasses } from "@/components/styles/container";
 
 import { PageHeader } from "../_components/PageHeader";
 
-import { AboutContent } from "./_components/AboutContent";
+import { DefiExpectations } from "./_components/DefiExpectations";
+import { PartnerCards } from "./_components/PartnerCards";
+import { InkTheFuture } from "./_components/InkTheFuture";
 
 export default function AboutPage() {
   const t = useTranslations("About");
   return (
     <div className={newLayoutContainerClasses()}>
       <PageHeader title={t("title")} description={t("description")} />
-      <AboutContent />
+      <InkTheFuture />
+      <DefiExpectations />
+      <PartnerCards />
     </div>
   );
 }
