+++ src/components/ParallaxedHoverImage.tsx
@@ -3,13 +3,19 @@ import Image, { ImageProps } from "next/image";
 
 import { classNames } from "@/util/classes";
 
-import { MagneticHoverEffect } from "./MagneticHoverEffect";
+import {
+  MagneticHoverEffect,
+  MagneticHoverEffectProps,
+} from "./MagneticHoverEffect";
 
-export interface ParallaxedHoverImageProps extends ImageProps {}
+export interface ParallaxedHoverImageProps extends ImageProps {
+  blendMode?: MagneticHoverEffectProps["blendMode"];
+}
 
-export const ParallaxedHoverImage: React.FC<ParallaxedHoverImageProps> = (
-  props
-) => {
+export const ParallaxedHoverImage: React.FC<ParallaxedHoverImageProps> = ({
+  blendMode,
+  ...props
+}) => {
   return (
     <MagneticHoverEffect
       easeType="ease"
@@ -17,6 +23,7 @@ export const ParallaxedHoverImage: React.FC<ParallaxedHoverImageProps> = (
       maxDistanceX={0.1}
       maxDistanceY={0.1}
       className="w-full h-full"
+      blendMode={blendMode}
     >
       <Image
         {...props}
